rewrite Node class for bfs
